requirements.txt
lark-parser==0.12.0
PySide6==6.9.3
PySide6_Addons==6.9.3
PySide6_Essentials==6.9.3
shiboken6==6.9.3
